Version 1

+++ app/src/main/java/com/example/a91902/androidble/database/MyDbHelper.java
@@ -9,7 +9,7 @@ import static com.example.a91902.androidble.database.Constants.*;
 public class MyDbHelper extends SQLiteOpenHelper {
     public static String DATABASE_NAME = "user";
     public static String DATABASE_CART = "cart";
-    public static int DATABASE_VERSION = 14;
+    public static int DATABASE_VERSION = 15;
     public static String DATABASE_USER = "CREATE TABLE " + TABLE_USER + "(" + ID + " TEXT, " + NAME + " TEXT, " + PHONE_NUMBER + " TEXT, " + EXPIRY + " TEXT, " + STATUS + " TEXT )";
     public static String DATABASE_CART_TABLE = "CREATE TABLE " + TABLE_CART + "(" + PRODUCT_NAME + " TEXT, " + PRODUCT_PRICE + " TEXT, " + PRODUCT_IMAGE + " INT )";
 
@@ -25,8 +25,11 @@ public class MyDbHelper extends SQLiteOpenHelper {
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        db.execSQL("DROP TABLE IF EXISTS "+TABLE_USER);
-        db.execSQL("DROP TABLE IF EXISTS "+TABLE_CART);
+        if (newVersion > oldVersion) {
+
+            db.execSQL("DROP TABLE IF EXISTS "+TABLE_USER);
+            db.execSQL("DROP TABLE IF EXISTS "+TABLE_CART);
+        }
         onCreate(db);
     }
 }
